Add -Wsign-compare flag and fix macro warnings

+++ misc/ta_talloc.h
@@ -115,7 +115,7 @@ char *ta_talloc_asprintf_append_buffer(char *s, const char *fmt, ...) TA_PRF(2, 
 #define MP_TARRAY_INSERT_AT(ctx, p, idxvar, at, ...)\
     do {                                            \
         size_t at_ = (at);                          \
-        mp_assert(at_ <= (idxvar));                 \
+        mp_assert(at_ <= (size_t)(idxvar));         \
         MP_TARRAY_GROW(ctx, p, idxvar);             \
         memmove((p) + at_ + 1, (p) + at_,           \
                 ((idxvar) - at_) * sizeof((p)[0])); \
@@ -130,7 +130,7 @@ char *ta_talloc_asprintf_append_buffer(char *s, const char *fmt, ...) TA_PRF(2, 
 #define MP_TARRAY_INSERT_N_AT(ctx, p, idxvar, at, c)\
     do {                                            \
         size_t at_ = (at);                          \
-        mp_assert(at_ <= (idxvar));                 \
+        mp_assert(at_ <= (size_t)(idxvar));         \
         size_t c_ = (c);                            \
         MP_TARRAY_GROW(ctx, p, (idxvar) + c_);      \
         memmove((p) + at_ + c_, (p) + at_,          \
@@ -143,7 +143,7 @@ char *ta_talloc_asprintf_append_buffer(char *s, const char *fmt, ...) TA_PRF(2, 
 #define MP_TARRAY_REMOVE_AT(p, idxvar, at)          \
     do {                                            \
         size_t at_ = (at);                          \
-        mp_assert(at_ <= (idxvar));                 \
+        mp_assert(at_ <= (size_t)(idxvar));         \
         memmove((p) + at_, (p) + at_ + 1,           \
                 ((idxvar) - at_ - 1) * sizeof((p)[0])); \
         (idxvar)--;                                 \
